wait for IFF_UP and IFF_RUNNING after calling ifup
This fixes https://bugzilla.redhat.com/show_bug.cgi?id=961184

Apparently one or the other of IFF_UP and IFF_RUNNING are not always
set by the time /sbin/ifup returns control to netcf, so the subsequent
check to verify that the interface is up may fail. This patch adds a
loop to re-check the status of the interface every 250msec for up to
2.5 seconds (or until both flags are set). If timeout is reached, it
still fails the operation.