support systemd-based netcf-transaction
https://bugzilla.redhat.com/show_bug.cgi?id=1044681

The netcf-transaction script is used in two ways:

1) It is called by libnetcf.so to begin, commit, and manually rollback
a set of network interface config changes.

2) It is called as part of normal system startup to automatically
rollback any config changes that are part of a netcf transaction that
hasn't yet been committed. (in other words, if you reboot after
calling ncf_change_begin() but before calling ncf_change_commit(), all
the changes you made subsequent to ncf_change_begin() will be erased).

This functionality is only supported for the "redhat" driver (used by
Fedora, CentOS, and RHEL), and up until now, it was all self-contained
in an initscripts-style shell script installed to
/etc/init.d/netcf-transaction - libnetcf.so would exec this script,
and it was also conveniently placed for initscripts to run it at
startup.

Some time ago, Fedora started supporting systemd for system
init/daemon management, but is still maintaining legacy support for
initscripts for packages that haven't yet moved over to systemd. RHEL7
also will support systemd.

This patch updates netcf to allow building packages either with an
initscripts-style script for /etc/init.d or a systemd-style init file
for /lib/systemd/system. The selection of which to include in the
build can either be made manually at configure-time (using the
--with-sysinit=(initscripts|systemd) option) or, if not explicitly
given on the configure commandline, configure.ac attempts to
automatically determine which style to use based on the existence of
the directory "/etc/systemd". Additionally, netcf.spec decides which
type of sysinit to build for based on the Fedora/RHEL version, so that
anyone building a netcf rpm on Fedora < 20 will continue to get
initscripts files rather than switching midstream to systemd.

The core functionality that was previously in
/etc/init.d/netcf-transaction has been moved to the file
/usr/libexec/netcf-transaction.sh, which is referenced both by the
initscripts init file (/etc/init.d/netcf-transaction) and by the
systemd service file (/lib/systemd/system/netcf-transaction.service):

  src/netcf-transaction.init.in (new initscripts file)
  src/netcf-transaction.service.in (systemd service file)
  src/netcf-transaction.sh.in (this was previously the initscripts file)

netcf.spec.in:
 * add check to decide systemd vs. initscripts based on
   RHEL/Fedora version and set configure commandline accordingly;

 * update %post and %postun to do the right thing for systems, and add
   a new %preun section for the same reason. Unlike libvirt, we know
   that any version of netcf chosen to use systemd by this specfile
   will already have the systemd rpm macros available, so we don't
   need the extra check for that which is present in the libvirt
   specfile. (this patch was written by heavily referencing the
   libvirt bits responsible for setting up the "libvirt-guests"
   service in libvirt).

src/Makefile.am:
 * generate the three new files when appropriate
 * change install-data-local and uninstall-local targets as appropriate
   for systemd vs initscripts vs none.

configure.ac:

 * support new --with-sysinit arg to manually specify
   systemd/initscripts/none.

 * autodetect which to use based on setting of $with_driver and
   presence of /etd/systemd directory.

NB: the location of the netcf-transaction script, which src/drv_*.c
find via the NETCF_TRANSACTION #define is currently hardcoded to look
in /usr/libexec. Ideally, this should use $libexecdir, but that
resolves to "${exec_prefix}/libexec" at configure time. To get the
desired results, we need to update gnulib and use gnulib's configmake
package, this way we'll end up with a NETCF_TRANSACTION that will be
set at the time make is run rather than configure. This is left as a
separate patch though, so that the initial support for systemd can be
backported to older releases of netcf (a gnulib update shouldn't be
backported to the maintenance branch of an already-cut release).