add enable_autotools to specfile
Another convenience borrowed from libvirt's build files. If a
downstream distro backports a patch that modifies configure.ac or one
of the Makefile.am's, the files in the source tarball will be out of
date and will need to be regerated during the build. This can now be
easily done by simply setting enable_autotools to 1 in the specfile.