Require bridge-utils for netcf-libs in the specfile
This resolves:

  https://bugzilla.redhat.com/show_bug.cgi?id=911403 (Fedora 20)
  https://bugzilla.redhat.com/show_bug.cgi?id=1060317 (RHEL7)

netcf itself doesn't use anything in bridge-utils, but it called
/sbin/ifup, which *can* run brctl (which is part of bridge-utils).
Technically, initscripts should require brutils (since /sbin/ifup is a
part of initscripts), but we all know that's not going to
happen. Instead we can just make netcf require bridge-utils. It's not
a very large package, so it's unlikely we'll get any complaints about
bloat.

(NB: in the past, libvirt had a Requires: bridge-utils, so this didn't
used to be a problem. libvirt removed this Requires: in v0.9.8, when
it switched from execing brctl to directly calling ioctls to create
and manage bridge devices.)