fix wrong status of bridge devices
The IFF_RUNNING flag is not set on a bridge device if there are no
devices attached to it. So bridges without attached devices are reported
as inactive. Destroying inactive devices leads to an error and thus the
device can't be removed (easily) from the operating system.

This patch checks if the device is a bridge and if so ignores the IFF_RUNNING
flag.