report link state/speed in interface status
This patch adds reporting of link state and speed for appropriate
types of interfaces. It is added to the interface status as a single
element:

  <link state='up|down|whatever' speed='nn'/>

(the speed is in Mbits/sec, i.e. exactly what is reported by the
kernel). The information come from the following locations in sysfs:

 state - /sys/class/net/$ifname/operstate
 speed - /sys/class/net/$ifname/speed

Note that speed will be set to 0 if operstate is anything but
"up". Also, this element is not added if an interface is of type
"bridge" (although any interfaces attached to the bridge, and reported
as subordinate interfaces in the same xml, *will* get a <link>
element.)

An example - this is a bridge device which has one bond attached; the
bond is comprised of two ethernets:

  <interface name="br0" type="bridge">
    <bridge>
      <interface name="bond0" type="bond">
        <link state="up" speed="1000"/>
        <bond>
          <interface name="p13p2_5" type="ethernet">
            <link state="up" speed="1000"/>
            <mac address="ea:8b:8d:18:67:fe"/>
          </interface>
          <interface name="p13p2_6" type="ethernet">
            <link state="up" speed="1000"/>
            <mac address="ea:8b:8d:18:67:fe"/>
          </interface>
        </bond>
      </interface>
    </bridge>
  </interface>

(note that in this case the bond reports a speed of 1000, although in
some cases I have seen it report "2000". This is an issue of the
kernel driver - netcf merely reports whatever the kernel has said.)