Fix netmask for prefix == 32
This resolves:

  https://bugzilla.redhat.com/show_bug.cgi?id=1116314

If an interface was defines with a network prefix of 32 (could be used
for a point-to-point link), instead of translating into a netmask of
255.255.255.255, it would end up being 0.0.0.0. This behavior was
caused by the function ipcalc_netmask() assuming that
~(0xffffffffu >> 32) would be 0xffffffffu, when in fact the result of
shifting right a 32 bit number by 32 positions is undefined. As it
happens, in the case of the code generated by gcc, the result of the
above is 0, when it should in fact be 0xffffffffu.

The safe way to deal with this is to simply special-case prefix == 32.