allow reordering of elements in interface xml
The interface xml schema was written with strict rules about the
ordering of the elements. This was never intentional, but just due to
omission of <interleave> in the appropriate places. This patch just
adds in <interleave> wherever there is more than one element, and
re-indents everything else appropriately.

NB: this change was earlier made to the libvirt interface.rng (in
upstream commit a341fc73, which was in libvirt-1.2.6. Also included in
libvirt-1.2.6 was commit 0b33d7c9, which re-orders the emited
interface XML; it turns out that a libvirt that contains the latter
commit (i.e. it re-orders the XML), when paired with a netcf that
doesn't have this current patch (allowing the re-ordering) will log
an error and fail any attempt to define a new host interface. Because
of this, once this patch is in a netcf release, we may want to
consider putting

  Requires: netcf >= 0.2.6

in the libvirt rpm for libvirt 1.2.6 and beyond.