permit link state in XML schema for bonds and vlans too
When support for reporting link state and speed in interface status
was added to netcf in commit 4be980b4, the changes to the RNG file
(XML schema) were taken from libvirt, but at that point libvirt's RNG
did not document reporting link state/speed for vlan or bond
interfaces, so the schema in netcf also only shows the presence of the
<link> element for ethernet interfaces (although in practice it does
actually report <link> for those types of interfaces).

This patch updates the RNG in netcf to match what is allowed by the
RNG in libvirt (since being corrected in libvirt commit
69db5f92). This makes no functional difference, since netcf's RNG is
only referenced when an interface is defined (but <link> is a
read-only element, so would never appear in the definition for a new
interface), but it's always a good idea for the schema to be correct.

libirt upstream commit 69db5f921a2f6d62645f3531042b13b6382eb865