Report file path and reason when aug_save fails
When aug_save fails, the /augeas metadata tree is now searched to find a
possible reason for the failure and an error is raised containing the file
path, Augeas error code (pointing to the action that failed) and the message if
supplied (from strerror).

A failure to unlink a file now results in a message such as:

  error: unspecified error
  error: aug_save failed on /etc/sysconfig/network-scripts/ifcfg-em1:
  unlink_orig (Permission denied)

Multiple failures aren't reported in detail, but will be printed if debug is
enabled.