Recognize IPADDR0/PREFIX0/NETMASK0/GATEWAY0 in redhat and suse backends
(another case of explanation taking longer than the patch...)

This resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1147650

The redhat and suse backends for netcf convert from a raw xml-ized
representation of ifcfg file contents (produced by augeas) to the
<interface> xml element known by libvirt using two xsl stylesheets:

  data/xml/redhat-put.xsl
  data/xml/suse-put.xsl

(I am personally interested in redhat-put.xsl, but am making the patch
described below to suse-put.xsl as well since the files are nearly
identical).

Examples of the "raw" xml can be seen in
tests/redhat/schema/*.xml. Basically each ifcfg setting is in an
element called "node", e.g:

  <node label="DEVICE" value="eth1"/>
  <node label="IPADDR" value="192.168.0.5"/>

The issue is that while the netcf stylesheets were written to use
IPADDR/PREFIX/NETMASK/GATEWAY, in certain versions of NetworkManager
(e.g. the version that is in RHEL7, but *not* the version that is in
Fedora 20) NetworkManager uses IPADDR0, PREFIX0, and GATEWAY0 to store
the interface IPv4 info, rather than IPADDR, PREFIX, and GATEWAY (In
those cases, it will still recognize IPADDR/PREFIX/GATEWAY, but when
it saves anything, it uses IPADDR0/PREFIX0/GATEWAY0).  Initscripts, on
the other hand, recognizes IPADDR0 and PREFIX0, but doesn't understand
GATEWAY0, it only sees GATEWAY.

The problem with the above comes when we do a dumpxml of an interface
created by NM (or anaconda, the Fedora installer, which also uses
IPADDR0), or worse yet when libvirt's virsh utility puts a bridge on
an ethernet by getting the current XML, modifying it, then writing
back the new config - since netcf doesn't see IPADDR0, the <ip>
element is empty in the original retrieved xml, and thus when the new
bridge interface is defined, it has no IP address info.

The solution to this is to improve the *-put.xsl to recognize either
the old form *or* the ${var}0 form when converting into <interface>
XML. For maximum compatibility, the conversion back into ifcfg format
is left the same, as all versions of initscripts and NetworkManager
that I've tried continue to recognize them.

Beyond what is done here, both NM and initscripts recognize IPADDRn
where "n" can be any number up to 255, and since we support multiple
IPv6 addresses in netcf, I would like to support multiple IPv4
addresses as well. That is left for a future patch, as my XSL skills
simply are not up to that task (patches welcome!)