build: apply patchfiles to rpm build dir using git
Rather than requiring a matching "patchXXX -p1" line in the specfile
for each "PatchXXX: blah.patch", this patch modifies the specfile so
that all that is needed is a list of patches ala "Patch XXX:
blah.patch", and uses git in a cleverly written (by Jirka Denemark
<jdenemar@redhat.com> for the libvirt specfile) script to apply any
patches found in the spec. Here is Jirka's comment from the commit log
in libvirt:

    With this change, any patch declared in libvirt.spec with Patch[0-9]* is
    automatically applied in %prep. Unlike with the standard %patch[0-9]*,
    patches are applied with "git am" to avoid some unexpected results.
    However, as a result of this, all patches must be in the right format
    for "git am" to be able to apply them; they should ideally be generated
    from git using "git format-patch".

It also has the advantage that when you "prep" a new source directory,
it is setup with a git tree so that you can easily create new patches
to add to the rpm (by "git commit"ing the changes, then "git
format-patch ..." to create the patchfiles).