build: fix/remove %install/%clean in specfile
The %clean target hasn't been need since Fedora13/RHEL6|CentOS6, so it
should be removed completely.

The cleaning up of the build directory prior to install has also been
done automatically for at least that long, so is not needed. *.la is
removed from the install image in $RPM_BUILD_ROOT because they are
unused on the target system.