when calling aug_get() escape special characters in components of paths
Use aug_escape_name_wrap() to escape any unquoted special characters
in a component of the path sent to aug_get(). This isn't needed when
the path being used is one returned from a call to aug_match() (those
strings are already escaped by augeas), or when the unsanitized data
is inside single quotes in an expression; after eliminating those two
cases and cases where the string is a constant, there is only a single
call to aug_get() left - in aug_get_mac().