Better messages on failure reading /sys/class/net/$dev/(operstate|speed)
add_link_info() sets an error message when it fails to read
/sys/class/net/$dev/operstate, but unfortunately frees the path string
just prior to setting the error message, so the error string has
"(null)" instead (thanks to glibc's asprintf noticing the null
pointer).

Additionally, the macro used to create the log message,
ERR_THROW_STRERROR() already turns errno into a string in the locally
defined errbuf, but we aren't including it in the format string, so it
doesn't get printed.

This patch makes sure the path isn't freed until it is really no
longer needed, and adds errbuf to the log message.

This will hopefully help to find the root cause of

  https://bugzilla.redhat.com/show_bug.cgi?id=1185850