Don't return error if /sys/class/net/$dev/operstate isn't found
This can happen for alias "interfaces", e.g: "eth0:1" (which aren't
really an interface, but are a deprecated method of adding multiple IP
addresses to a single interface). Since only the base interface
("eth0") is really a device, only it has an entry in /sys/class/net.

It's not important that such interfaces have truly meaningful
information in their <link> element (as we don't really have any truly
meaningful state, and haven't historically provided other status for
alias interfaces (e.g. IP address). What *is* important is that we
don't return an error, as this breaks applications that rely on being
able to get back a successful (if empty) status for any interface
listed.

This resolves the regression with virt-manager described here:

  https://bugzilla.redhat.com/show_bug.cgi?id=1185850