local copy of new augeas function aug_escape_name()
In order to get the functionality of aug_escape_name() on systems that
use an older version of augeas that doesn't yet have this API, this
patch replicates the same functionality in a local function
aug_escape_name_base() (which will be called by aug_escape_name_wrap()
in most places, or directly when augeas isn't available, e.g. during
xslt transforms).

Once a release of augeas with the new API is generally available, a
followup patch will selectively replace the innards of
aug_escape_name_wrap() with a call to the real aug_escape_name(), if
configure detects it.

Note that this patch by itself does not solve CVE-2014-8119. A full
solution requires augeas itself to be patched to change the behavior
of aug_match() so that it escapes all the components of returned
paths.