check for valid pointer on all returns from aug_get()
Several calls to aug_get() assumed that if the return value wasn't <0,
that it was returning a valid pointer to a string. Aside from the fact
that a return of 0 indicates no string, there is also the problem that
even if the return value is 1, the string could be NULL.

In most cases this didn't cause a problem in practice, because the
path being sent to aug_get() was a path previously returned by
aug_match() (implying that it really does point to a value).