Allow IPv4 static addresses simultaneous with DHCP
Because initscripts network startup ends up ignoring static IPv4
addresses when dhcp is enabled (BOOTPROTO=dhcp in the ifcfg file) and
because at least some of the frontends to Network Manager disable
configuring a static address if dhcp is enabled (and likely due to
other indications, but this all happened far in the past...) netcf has
always made DHCP and static IP addresses for IPv4 mutually exclusive
in the config - an interface either had one or it had the other.

A recent BZ filed against libvirt:

  https://bugzilla.redhat.com/show_bug.cgi?id=1113983

suggested that we should also forbid simultaneous DHCP and static IPs
for IPv6, but investigation shows the following:

1) there is nothing inherent in the DHCP protocol that precludes
having both a single DHCP address and one or more static IP addresses
from each protocol (as well as an IPv6 address learned via IPv6
autoconf). Even the lead NetworkManager developer says this in an
email response to a question posed on the dnsmasq mailing list:

  http://lists.thekelleys.org.uk/pipermail/dnsmasq-discuss/2015q1/009346.html

2) Different networking backends currently properly support different
combinations of these address types for the two platforms (for
example, NetworkManager 0.9.10.2 on Fedora 21 will properly acquire a
DHCPv4 address to live alongside multiple static IPv4 addresses while
initscripts 9.56.7 will discard the static IPs, and initscripts will
support multiple static IPv6 addresses together with an autoconf
address, while NetworkManager apparently won't (although this could be
a configuration error on my part).

In the end, any time that all three address types for IPv6 (autoconf,
dhcp, static) or two address types for IPv4 (dhcp, static) cannot work
simultaneously on the same interface, that is a bug in the networking
backend (e.g. initscripts, NetworkManager), *not* a reason to disallow
such configurations via netcf.

For that reason, this patch opens up the RNG for the IPv4 protocol,
allowing <dhcp/> to be specified simultaneously with any number of
static <ip> elements for IPv4, just as we already did for IPv6. Any
future failures of such configurations to work properly should result
in bugs filed against the network backend in use at the time (or
possibly the dhcp server, or some other piece of infrastructure).