dutil: Untabify, to pacify gcc
dutil.c: In function 'rng_validate':
dutil.c:324:5: warning: this 'if' clause does not guard... [-Wmisleading-indentation]
     if (r != 0 && ncf->errcode == NETCF_NOERROR)
     ^~
dutil.c:328:2: note: ...this statement, but the latter is misleadingly indented as if it is guarded by the 'if'
  xmlRelaxNGFreeValidCtxt(ctxt);
  ^~~~~~~~~~~~~~~~~~~~~~~

The warning is caused because the 'if' was indented with four spaces,
while the 'xmlRelaxNGFreeValidCtxt' was indented with TAB; and gcc
couldn't figure out that we were asking for TAB to represent 4 instead
of 8 spaces.  Rather than add more TABs, I'd rather nuke them all.

Patch best viewed with 'git diff -b'.

Signed-off-by: Eric Blake <eblake@redhat.com>