slience a false Coverity report
Coverity complains that

   unref(nif, netcf_if);

sets nif to NULL without free'ing the memory it points to. But unref()
is a macro (defined in src/ref.h) that uses a refcount in the object
to automatically free it when there are no more pointers to it.

Since the code is correct, we add a comment that silences the Coverity warning

Signed-off-by: Laine Stump <laine@laine.org>